devices classes cleaned and reworked a little bit
