+++ src/UIA.Framework/Configuration/Timeouts.cs
@@ -3,5 +3,6 @@
     public class Timeouts
     {
         public const int Action = 100;
+        public const int Device = 100;
     }
 }
