--- src/UIA.Framework/Devices/Commands/KeyboardCommands.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace UIA.Framework.Devices.Commands
-{
-    [Flags]
-    internal enum KeyboardCommands
-    {
-        ExtendedKey = 0x0001,
-        KeyUp = 0x0002
-    }
-}
