+++ src/UIA.Framework/Devices/Mouse.cs
@@ -1,67 +1,57 @@
 ﻿using System;
-using System.Drawing;
 using System.Runtime.InteropServices;
-using System.Windows.Forms;
-using UIA.Framework.Devices.Commands;
+using UIA.Framework.Configuration;
+using UIA.Framework.Helpers;
 
 namespace UIA.Framework.Devices
 {
-    public static class Mouse
+    public class Mouse
     {
-        private static Point screen = new Point(Screen.PrimaryScreen.Bounds.Width, Screen.PrimaryScreen.Bounds.Height);
-
-        [StructLayout(LayoutKind.Sequential)]
-        public struct CursorInfo
+        [Flags]
+        private enum MouseCommands
         {
-            public uint size;
-            public uint flags;
-            public IntPtr handle;
-            public Point point;
-
-            public static CursorInfo New()
-            {
-                return new CursorInfo { size = (uint)Marshal.SizeOf(typeof(CursorInfo)) };
-            }
+            Absolute = 0x8000,
+            LeftDown = 0x0002,
+            LeftUp = 0x0004,
+            MiddleDown = 0x0020,
+            MiddleUp = 0x0040,
+            Move = 0x0001,
+            RightDown = 0x0008,
+            RightUp = 0x0010,
+            WheelMove = 0x0800,
+            XDown = 0x0080,
+            XUp = 0x0100,
+            Wheel = 0x0800,
+            HWheel = 0x01000
         }
 
-        enum SystemMetric
+        private enum SystemMetric
         {
             SM_CXSCREEN = 0,
             SM_CYSCREEN = 1,
         }
 
-        [DllImport("user32.dll")]
-        static extern bool GetCursorInfo(ref CursorInfo cursorInfo);
-
         [DllImport("user32.dll", EntryPoint = "mouse_event")]
-        static extern void MouseEvent(int dwFlags, int dx, int dy, int dwData, int dwExtraInfo);
+        private static extern void MouseEvent(int dwFlags, int dx, int dy, int dwData, int dwExtraInfo);
 
         [DllImport("user32.dll")]
-        static extern bool GetCursorPos(ref System.Drawing.Point cursorInfo);
+        private static extern int GetSystemMetrics(SystemMetric smIndex);
 
-        [DllImport("user32.dll", EntryPoint = "SetCursorPos")]
-        private static extern bool SetCursorPos(int x, int y);
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        static extern int GetLastError();
-
-        [DllImport("user32.dll")]
-        static extern int GetSystemMetrics(SystemMetric smIndex);
-
-        static int CalculateAbsoluteCoordinateX(int x)
+        private static int CalculateAbsoluteCoordinateX(int x)
         {
             return (x * 65536) / GetSystemMetrics(SystemMetric.SM_CXSCREEN);
         }
 
-        static int CalculateAbsoluteCoordinateY(int y)
+        private static int CalculateAbsoluteCoordinateY(int y)
         {
             return (y * 65536) / GetSystemMetrics(SystemMetric.SM_CYSCREEN);
         }
 
         public static void Click(double x, double y)
         {
-            var action = (int)MouseCommands.Absolute | (int)MouseCommands.Move | (int)MouseCommands.LeftDown | (int)MouseCommands.LeftUp;
-            MouseEvent(action, CalculateAbsoluteCoordinateX((int)x), CalculateAbsoluteCoordinateY((int)y), 0, 0);
+            MouseEvent((int)MouseCommands.Absolute | (int)MouseCommands.Move, CalculateAbsoluteCoordinateX((int)x), CalculateAbsoluteCoordinateY((int)y), 0, 0);
+            Wait.Milliseconds(Timeouts.Device);
+            MouseEvent((int)MouseCommands.LeftDown | (int)MouseCommands.LeftUp, 0, 0, 0, 0);
         }
     }
 }
