+++ src/UIA.Framework/UIA.Framework.csproj
@@ -51,10 +51,8 @@
     <Compile Include="Application\Application.cs" />
     <Compile Include="Configuration\ActionHandler.cs" />
     <Compile Include="Configuration\Timeouts.cs" />
-    <Compile Include="Devices\Commands\KeyboardCommands.cs" />
-    <Compile Include="Devices\Commands\VirtualKeys.cs" />
-    <Compile Include="Devices\Commands\MouseCommands.cs" />
-    <Compile Include="Devices\Inputs\MouseCursor.cs" />
+    <Compile Include="Devices\Keyboard.cs" />
+    <Compile Include="Devices\Cursor.cs" />
     <Compile Include="Devices\Mouse.cs" />
     <Compile Include="Elements\Attributes\DependsAttribute.cs" />
     <Compile Include="Elements\Attributes\NeverAttribute.cs" />
